Switch clips to storing data in sketchpad, add helpers, adjust users
We also introduce the originalPath property on clip metadata, which is
set explicitly when you import a wave file into a clip. This is then
used by SampleLibrary to re-select the same entry on your system when
you select the slot. If the file is not found, nothing happens, unless
the folder it would have been contained within is (at which point that
folder will be selected)