sketchpad.song: Fix missing sketchFx eq persistence, move passthrough serialiser function
The first is potentially going to not matter a huge deal at some point,
if we end up wanting to hide the loop option, but even if we do, we'll
also eventually want to return it, so... fix it.

Also, the passthrough serialiser wants to be accessible from outside of
the class, so put it somewhere it's... accessible.