gui.layer: Ensure that mixer and eq settings are persisted in the snapshots
Without this, we end up, well, without the ability to pass that
information around when using snapshot data (which directly affects our
.snd files, bouncing and unbouncing, and all that sort of thing). The
per-slot pan as yet not exposed in the ui, but let's just store it
anyway, thinking we'll likely want to add that at some point. Need to
work out where it'd want to live, though.