zyngine: Switch to using the explicit MidiRouter calls for feedback and engine cc writes
This has been a problem for some time, and was base on some incorrect
assumptions about what should happen when. So, we now have a way to
write explicitly to a single synth engine's output, and also a way to
handle controller feedback without affecting things that are also just
there. This should hopefully make a few things at least a little more
sane.