engine.fluidsynth: Fix fetching presets with non-loaded bank, and correct stripping
The first issue was what was causing the system to attempt to fetch
presets for an instrument called "True", rather than the number it was
supposed to be sending out. It should also help just in general, as the
previous logic was kind of wonky anyway.

The second part is just because we would previously snip off the last
letter in some cases, and that should no longer happen (trust the
stripping code, instead of trying to do our own bit of odd magic).