sketchpad.channel: Save snapshot when passthrough changes, clear passthrough fully when clearing slots
When changing values on a passthrough client (the synth and fx ones in
particular), we no longer need to ask the song to save itself, as this
is a part of the *snapshot*, not the song. So, don't do that. Instead,
save the snapshot.

When clearing a slot, the passthrough values should be cleared. That
also means the snapshot should be saved, so let's make sure that
happens.