sketchpad.channel: Move the old keyzone auto logic here from qml
The previous home for this required the channel to be the current one
for keyzoning changes to be applied. This would usually be true anyway,
but it just was very extremely the wrong place for it anyway. Also, this
should further ensure that when swapping slots, the keyzones are updated
as well.