selectedSlot logic: Rework to be less magic, and also move a lot out of qml
Previously, the logic for the selected slots on tracks was kind of all
over the place. In particular, it was in qml, which really isn't where
that belongs (since it's not *tracked* there, the initial state logic
really wants to be elsewhere), and so, it now lives in the track itself,
where we can control it a bit more tightly.

This also removes the use of the component stored in the selected slot
object, which while still useful for things like the lastSelectedObject
stuff, just kind of got in the way for selectedSlot.

Things feel much more understandable now, and while we can perhaps work
out some more fun things to work on here, this will do very nicely for
us for now.

The pickFirstAndBest logic is still used for, specifically, the keyzone
setup, where it really is quite important that we've got one of the
synth or sample slots selected for it to make some semblance of sense.

This also makes the "wait, should i activate the thing" clicky tappy
logic stuff much more straightforward.