gui.sketchpad: Remove redundant clipsToRecord prop, minor rework of recording enqueue function
The removal of the clipsToRecord property is basically because it was
some really old logic that didn't really make any sense the way we do
things now.

The minor reworking of the clip recording enqueueing function is to
ensure that if we bail out, we leave things in a more untouched state.
Also, we now handle sources based on filled sound slots, not explicitly
on synth layers (because we can also record things which have samples in
them only, and we would previously be told we could not).