sketchpad.channel/sounds: Add track style property, handle for sounds
This one property is intended to cover the styles we have for our
tracks, as a sort of "do a bunch of things at the same time" system,
which lets us much more easily handle those properties together, without
losing the ability to in the future add in more abilities. This also
means adding in storing and restoring this when dealing with the sounds,
as this fundamentally defines how the sound of a track works.