TracksBar/Settings: Add style toggle (and handle option visibility for style)
When in anything other than Manual style, we need to be hiding the
routing style, slot picking style, and keyzone setup controls. We
currently just switch between everything and one-to-one mode, so it's
just a toggle for now, but i'm feeling like maybe this is not
sufficiently descriptive of what it's actually doing... maybe a quick
little popup in the style of the slot picking style selector would do
well for this purpose.