sketchpad.clip: Add a title property, use for the slot display
For now this will just return whatever the filename is, minus the suffix
(including a couple of the double ones, specificaly .sketch.wav and
clip.wav), so that we can see just a bit more of the actual filename in
the slots. It is not persisted, but it would be very useful to be able
to just... be able to edit this title as we might desire.