Sounds: Initial implementation for slot pull logic
This seems to pretty much be doing what is wanted, but it looks like
perhaps the synth snapshots don't work quite right (or the sound i'm
using to test is outdated or broken somehow). Pushing this, though, as
it's working as intended. Perhaps some polishing up on the display logic
would be good maybe.