Sounds: Fix type error on switching (and just ensure the logic applies)
This ensure that when we switch, the filter actually *is* reloaded,
whereas previously it would attempt to set it to the same value it
already was, which would result logically in the call being ignored
(because it would detect it was the same and do nothing), but also the
old code caused a type error, and this doesn't.

Also minor adjustment to the wording in the slots popup thing.