stepsequencer: Rejig layout, entirely remove pattern popup
Since the functionality previously found in the pattern popup now all
exists in the new track/clip picker location, kind of makes sense to
just... get rid of that. a *huge* pile of unneeded and overly intricate
code, that.

Also, rejig the layout so the track/clip picker now seems to simply
replace the steps, and put the settings row above them, so they cover
the bottom row of the note grid.