stepsequencer: Rework the layout (and light interaction) for track/clip picker
Documentation getting updated shortly, but in short: The track/clip
picker has moved up to the top.

If the track picker is open, bk changes the current track
If the settings panel is open, bk changes the current bar, unless track
picker is open.
If both are open, the top row of the notes grid is hidden, so it doesn't
get too cramped.
If neither is open, or notesettings is, bk changes the selection