sketchpad.clip: Explicitly request the old CAS is deleted
Otherwise we end up sort of... with a bunch of these things hanging
about, which isn't great.

It also fixes the problem where loading a sample into a slot where there
already is one loaded will keep playing the old sample (in particular in
1:1 mode), which of course is quite silly. The above description is the
larger issue, but audibly, this is, so... both? ;)