autoconnect: Hugely simplify the fx connection logic, and implement dynamic port creation
The first removes the old logic which was a relic of when we had cyclic
connections through the track passthrough when in standard mode, whereas
now we instead have an fx passthrough per-fx, which allows for much,
much simpler routing. We also now instead run the fx into the wet ports
on the passthrough, and the dry signal into the passthrough's dry port
as well as the fx, so we can do proper dry/wet mixing, and also actually
apply the eq/comp to the mixed-down signal, instead of the signal fed to
the fx itself (in other words: it now works as one would reasonably
expect given how the ui is laid out).

The second part ensures that we only have the ports in our graph that
are actually used. This hugely simplifies the layout (which is great for
development purposes), and more importantly also greatly decreases the
processing cost, since the ports which don't exist aren't causing
slowdowns any longer.

Much simpler, much nicer in general. Liking this.