Sketchpad Main: Add eq control and switcher/menu buttons for sends and eq/comp
This is the first iteration of displaying and controlling the eq from
the mixer display. Tap the eq thumbnail to switch between active bands,
and swipe to move that band around.

For the sidebar buttons, tap to switch, or tap the active button to show
a popup (not implemented for sends, that'll be a menu of some
description, for selecting which sends to use, and perhaps a couple of
other nice bits). For eq, this opens the SlotEqualizer dialogue for the
current track.