Sketchpad mixer EQ: Sort out the tap/slide interaction for the eq thumbnails
Now, when you tap an eq thumbnail, if the track it's for is not active,
that track is activated, and if the track is already the current track,
we cycle through the active eq bands. Swiping the finger across the
thumbnail will, whether that eq's track is active or not, move the
selected eq band (and also ensure there's an active one selected, and
if there's none active, it will not change the curves).