GlobalSeq: step+bk (and bar+bk, and clip+bk) transposes step entries
For step, it transposes the step's entries. For bar, it transposes all
entries in the given bar. For clip, it transposes the entire clip.

When transposing a clip, we further spit out the name of whatever is in
there. This likely will require just a touch of work to be done on the
Chords functionality, so we don't spit out a shorthand that's just "some
note" when we don't know what's in there. Just a list of all the notes
in that case would work.